pyproject.toml
[project]
name = "aws-safe-mcp"
version = "0.1.0"
description = "Read-only, opinionated MCP server for safely investigating AWS resources."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Haresh Patel" }
]
keywords = ["aws", "mcp", "model-context-protocol", "lambda", "serverless"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Debuggers",
  "Topic :: System :: Systems Administration",
]
dependencies = [
  "boto3>=1.34.0",
  "botocore[crt]>=1.34.0",
  "mcp>=1.2.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
]

[project.urls]
Homepage = "https://github.com/harryhazza77/aws-safe-mcp"
Repository = "https://github.com/harryhazza77/aws-safe-mcp"
Issues = "https://github.com/harryhazza77/aws-safe-mcp/issues"
Changelog = "https://github.com/harryhazza77/aws-safe-mcp/blob/main/CHANGELOG.md"
Security = "https://github.com/harryhazza77/aws-safe-mcp/security/policy"

[project.scripts]
aws-safe-mcp = "aws_safe_mcp.main:main"

[dependency-groups]
dev = [
  "bandit>=1.7.10",
  "pip-audit>=2.7.3",
  "pytest>=8.0.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.6.0",
  "mypy>=1.10.0",
  "types-PyYAML>=6.0.12.20240311",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aws_safe_mcp"]

[tool.hatch.build.targets.sdist]
include = [
  "CHANGELOG.md",
  "CONTRIBUTING.md",
  "LICENSE",
  "README.md",
  "SECURITY.md",
  "docs",
  "examples",
  "src",
  "tests",
  "pyproject.toml",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
branch = true
source = ["aws_safe_mcp"]

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_also = [
  "if TYPE_CHECKING:",
  "if __name__ == .__main__.:",
  "raise NotImplementedError",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.bandit]
exclude_dirs = ["tests"]
skips = ["B101"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
packages = ["aws_safe_mcp"]

[[tool.mypy.overrides]]
module = ["boto3", "botocore.*"]
ignore_missing_imports = true
